refactor(web): extract the passkey sign-in block from the sign-in page
Master's OAuth continuation work grew SignInPage and this branch's
entry-point link pushed it past the 300-line component limit. The
passkey block is a self-contained widget — its own action, its own
conditional-UI preload effect, its own error surface — so it moves to
components/auth/passkey-sign-in.tsx beside the other auth widgets.
SignInPage forwards its signInPasskey port, so the page's test contract
is unchanged.
